+++ applications/main/nfc/helpers/nfc_detected_protocols.c
@@ -45,7 +45,7 @@ void nfc_detected_protocols_set(
     furi_assert(types);
     furi_assert(count < NfcProtocolNum);
 
-    memcpy(instance->protocols_detected, types, count);
+    memcpy(instance->protocols_detected, types, count * sizeof(NfcProtocol));
     instance->protocols_detected_num = count;
     instance->selected_idx = 0;
 }