+++ applications/main/nfc/helpers/protocol_support/mf_classic/mf_classic_extra_scenes.c
@@ -1,12 +1,10 @@
 #include "mf_classic_extra_scenes.h"
 
 #include "nfc/nfc_app_i.h"
-#include "../nfc_protocol_support_gui_common.h"
 #include <bit_lib/bit_lib.h>
 #include <dolphin/dolphin.h>
 #include <toolbox/stream/buffered_file_stream.h>
 #include <nfc/protocols/mf_classic/mf_classic_listener.h>
-#include "loader/loader.h"
 #include <nfc/protocols/mf_classic/mf_classic_poller.h>
 
 enum {
@@ -114,7 +112,6 @@ static NfcCommand nfc_dict_attack_worker_callback(NfcGenericEvent event, void* c
                 }
             }
         } else {
-            // Standard dictionary: keys are sizeof(MfClassicKey) bytes
             if(keys_dict_get_next_key(
                    instance->nfc_dict_context.dict, key.data, sizeof(MfClassicKey))) {
                 key_found = true;
@@ -412,6 +409,7 @@ static bool mf_classic_scene_dict_attack_on_event(NfcApp* instance, SceneManager
             if(state == DictAttackStateCUIDDictInProgress) {
                 nfc_poller_stop(instance->poller);
                 nfc_poller_free(instance->poller);
+                instance->poller = NULL;
                 keys_dict_free(instance->nfc_dict_context.dict);
                 if(instance->nfc_dict_context.cuid_key_indices_bitmap) {
                     free(instance->nfc_dict_context.cuid_key_indices_bitmap);
@@ -427,6 +425,7 @@ static bool mf_classic_scene_dict_attack_on_event(NfcApp* instance, SceneManager
             } else if(state == DictAttackStateUserDictInProgress && !(ran_nested_dict)) {
                 nfc_poller_stop(instance->poller);
                 nfc_poller_free(instance->poller);
+                instance->poller = NULL;
                 keys_dict_free(instance->nfc_dict_context.dict);
                 scene_manager_set_scene_state(
                     instance->scene_manager,
@@ -464,6 +463,7 @@ static bool mf_classic_scene_dict_attack_on_event(NfcApp* instance, SceneManager
                 if(instance->nfc_dict_context.is_card_present) {
                     nfc_poller_stop(instance->poller);
                     nfc_poller_free(instance->poller);
+                    instance->poller = NULL;
                     keys_dict_free(instance->nfc_dict_context.dict);
                     if(instance->nfc_dict_context.cuid_key_indices_bitmap) {
                         free(instance->nfc_dict_context.cuid_key_indices_bitmap);
@@ -485,6 +485,7 @@ static bool mf_classic_scene_dict_attack_on_event(NfcApp* instance, SceneManager
                 if(instance->nfc_dict_context.is_card_present) {
                     nfc_poller_stop(instance->poller);
                     nfc_poller_free(instance->poller);
+                    instance->poller = NULL;
                     keys_dict_free(instance->nfc_dict_context.dict);
                     scene_manager_set_scene_state(
                         instance->scene_manager,
@@ -515,6 +516,7 @@ static bool mf_classic_scene_dict_attack_on_event(NfcApp* instance, SceneManager
 static void mf_classic_scene_dict_attack_on_exit(NfcApp* instance) {
     nfc_poller_stop(instance->poller);
     nfc_poller_free(instance->poller);
+    instance->poller = NULL;
 
     dict_attack_reset(instance->dict_attack);
     scene_manager_set_scene_state(
@@ -558,7 +560,6 @@ static void mf_classic_scene_dict_attack_on_exit(NfcApp* instance) {
 }
 
 // ---- detect_reader -----------------------------------------------------
-#undef TAG
 #define NXP_MANUFACTURER_ID (0x04)
 
 #define NFC_SCENE_DETECT_READER_PAIR_NONCES_MAX        (10U)
@@ -712,7 +713,6 @@ static void mf_classic_scene_detect_reader_on_exit(NfcApp* instance) {
 }
 
 // ---- mfkey_complete ----------------------------------------------------
-#undef TAG
 typedef enum {
     NfcSceneMfClassicMfKeyCompleteStateAppMissing,
     NfcSceneMfClassicMfKeyCompleteStateAppPresent,
@@ -801,7 +801,6 @@ static void mf_classic_scene_mfkey_complete_on_exit(NfcApp* instance) {
 }
 
 // ---- mfkey_nonces_info -------------------------------------------------
-#undef TAG
 static void mf_classic_scene_mfkey_nonces_info_callback(
     GuiButtonType result,
     InputType type,
@@ -1134,6 +1133,7 @@ static bool mf_classic_scene_update_initial_on_event(NfcApp* instance, SceneMana
 static void mf_classic_scene_update_initial_on_exit(NfcApp* instance) {
     nfc_poller_stop(instance->poller);
     nfc_poller_free(instance->poller);
+    instance->poller = NULL;
 
     scene_manager_set_scene_state(
         instance->scene_manager,